svn merge -r 2782:2786 http://code.taobao.org/svn/tfs/branches/stable-2.5 ./trunk http://rb.corp.taobao.com/r/84443
git-svn-id: http://code.taobao.org/svn/tfs@2814 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ trunk/src/life_cycle_manager/expire_server/clean_task_helper.cpp
@@ -19,6 +19,7 @@
 #include <malloc.h>
 #include "common/func.h"
 #include "common/tairengine_helper.h"
+#include "common/atomic.h"
 
 using namespace std;
 namespace tfs
@@ -31,6 +32,7 @@ namespace tfs
     const int32_t SCAN_LIMIT = 1000;
     const int32_t SEC_ONE_DAY = 24 * 60 * 60;
     const int32_t MAX_INT32 = (1<<31) - 1;
+    const char del = '/';
 
     enum
     {
@@ -43,6 +45,7 @@ namespace tfs
     {
       kv_engine_helper_ = NULL;
       tair_lifecycle_area_ = 0;
+      running_task_count_ = 0;
     }
 
     CleanTaskHelper::~CleanTaskHelper()
@@ -63,9 +66,16 @@ namespace tfs
         kv_engine_helper_ = kv_engine_helper;
       }
 
+      if (TFS_SUCCESS == ret)
+      {
+        ret = restful_client_.initialize(SYSPARAM_EXPIRESERVER.nginx_root_.c_str(),
+                                                   SYSPARAM_EXPIRESERVER.es_appkey_.c_str());
+        TBSYS_LOG(INFO, "ret: %d, restful_client initialize SUCCESS", ret);
+      }
+
       tair_lifecycle_area_ = SYSPARAM_EXPIRESERVER.lifecycle_area_;
 
-      clean_task_state_ = 0;
+      running_task_count_ = 0;
 
       if (TFS_SUCCESS ==  ret)
       {
@@ -81,13 +91,14 @@ namespace tfs
       return ret;
     }
 
-    int32_t CleanTaskHelper::get_task_state()
+    int32_t CleanTaskHelper::get_running_threads_count()
     {
-      return clean_task_state_;
+      return running_task_count_;
     }
 
 
-    int CleanTaskHelper::take_note(const uint64_t local_ipport, const int32_t num_es,
+    //record stat to kv store
+    int CleanTaskHelper::stat_to_kvstore(const uint64_t local_ipport, const int32_t num_es,
                                    const int32_t task_time, const int32_t hash_bucket_num,
                                    const int64_t sum_file_num)
     {
@@ -139,9 +150,10 @@ namespace tfs
     }
 
 
-    int CleanTaskHelper::clean_task(const uint64_t local_ipport, const int32_t total_es, const int32_t num_es,
+    int CleanTaskHelper::do_clean_task(const uint64_t local_ipport, const int32_t total_es, const int32_t num_es,
                                     const int32_t note_interval, const int32_t task_time)
     {
+      UNUSED(local_ipport);
       int32_t ret;
       int32_t days_secs;
       int32_t hours_secs;
@@ -153,8 +165,18 @@ namespace tfs
       int32_t end_bucket_num;
       std::string file_name;
 
-      clean_task_state_ = 1;
+      struct timeval start;
+      struct timeval end;
 
+      struct timeval scan_start;
+      struct timeval scan_end;
+      int32_t g_scan_time = 0;
+
+      gettimeofday(&start, NULL);
+
+      int32_t expire_file_count = 0;
+
+      atomic_inc(&running_task_count_);
 
       ret = ExpireDefine::transfer_time(task_time, &days_secs, &hours_secs);
 
@@ -194,7 +216,6 @@ namespace tfs
           int64_t first, second;
           first = Func::get_monotonic_time();
           zero_secs = 0;
-
 
           for (hash_mod = start_bucket_num; hash_mod <= end_bucket_num; ++hash_mod)
           {
@@ -243,16 +264,24 @@ namespace tfs
                 while (go_on)
                 {
                   int32_t result_size = 0;
+                  gettimeofday(&scan_start, NULL);
                   ret = kv_engine_helper_->scan_keys(tair_lifecycle_area_, start_key, end_key, SCAN_LIMIT, first,
                                                      &kv_value_keys, &kv_value_values,
                                                      &result_size, scan_type);
+
+                  gettimeofday(&scan_end, NULL);
+
+                  g_scan_time += 1000000 * (scan_end.tv_sec - scan_start.tv_sec) + scan_end.tv_usec - scan_start.tv_usec;
+
                   if (EXIT_KV_RETURN_DATA_NOT_EXIST == ret)
                   {//no data
-                    TBSYS_LOG(ERROR, "data not exist");
+                    TBSYS_LOG(DEBUG, "data not exist, result_size: %d", result_size);
                     ret = TFS_SUCCESS;
                   }
 
-                  for(i = 0; i < result_size; ++i)
+                  expire_file_count += result_size;
+
+                  for (i = 0; i < result_size; ++i)
                   {
                     ret = ExpireDefine::deserialize_exptime_app_key(kv_value_keys[i]->get_data(),
                                                  kv_value_keys[i]->get_size(),
@@ -268,22 +297,72 @@ namespace tfs
                     if ((int32_t)(second - first) >= note_interval)
                     {
                        first = second;
-                       take_note(local_ipport, num_es, task_time, hash_mod, sum_file_num);
+                       //stat_to_kvstore(local_ipport, num_es, task_time, hash_mod, sum_file_num);
+                       TBSYS_LOG(INFO, "task num:%d this tasktime:%d start_bucket_num:%d end_bucket_num:%d now hash_num:%d now deal file: %s", num_es, task_time, start_bucket_num, end_bucket_num, hash_mod, t_file_name.c_str());
                     }
 
+                    uint32_t file_len = t_file_name.length();
+                    if (file_len > 0 && (t_file_name[0] == 'T' || t_file_name[0] == 'L'))
+                    {
+                      ret = restful_client_.unlink(t_file_name.c_str(), NULL, tfs::restful::DELETE, t_appkey.c_str());
+                    }
+                    else if (file_len > 0)
+                    {
+                      //file_name like appid/uid/file/aa/bb
+                      string::size_type pos = 0;
+                      string::size_type uid_pos = string::npos;
+                      string::size_type metaname_pos = string::npos;
 
+                      int i = 0;
+                      TBSYS_LOG(INFO, "file_name: %s", t_file_name.c_str());
+                      for (i = 0; i < 3; ++i)
+                      {
+                        pos = t_file_name.find_first_of(del, pos);
+                        if (pos != string::npos)
+                        {
+                          if (i == 0)
+                          {
+                            uid_pos = pos;
+                          }
+                          else if (i == 2)
+                          {
+                            metaname_pos = pos;
+                            break;
+                          }
+                          pos++;
+                        }
+                        else
+                        {
+                          TBSYS_LOG(ERROR, "file_name: %s is illegal", t_file_name.c_str());
+                          break;
+                        }
+                      }
 
-                    restful_client_.unlink(t_file_name.c_str(), NULL, tfs::restful::DELETE ,t_appkey.c_str());
-
-                  }
+                      if (uid_pos != string::npos && metaname_pos != string::npos)
+                      {
+                        int64_t appid = atoll(t_file_name.c_str());
+                        int64_t uid = atoll(t_file_name.c_str() + uid_pos + 1);
+                        //string meta_name = t_file_name.substr(metaname_pos, file_len - metaname_pos);
+                        TBSYS_LOG(INFO, "appid: %"PRI64_PREFIX"d, uid: %"PRI64_PREFIX"d, meta_name: %s",
+                                  appid, uid, t_file_name.c_str() + metaname_pos);
+                        if (appid < 0 || uid < 0)
+                        {
+                            TBSYS_LOG(ERROR, "error appid: %"PRI64_PREFIX"d, uid: %"PRI64_PREFIX"d, meta_name: %s",
+                                      appid, uid, t_file_name.c_str() + metaname_pos);
+                            continue;
+                        }
+                        ret = restful_client_.rm_file(uid, t_file_name.c_str() + metaname_pos, t_appkey.c_str(), appid);
+                      }
+                    }//else if() custom
+                  }//for
 
                   TBSYS_LOG(DEBUG, "this time result_size is: %d", result_size);
 
-                  if(result_size == SCAN_LIMIT)
+                  if (result_size == SCAN_LIMIT)
                   {
                     first = 1;
                     ret = ExpireDefine::serialize_exptime_app_key(days_secs, hours_secs, hash_mod, t_file_type,
-                          t_file_name, &start_key, start_key_buff, KEY_BUFF_SIZE);
+                        t_file_name, &start_key, start_key_buff, KEY_BUFF_SIZE);
                   }
                   else
                   {
@@ -313,10 +392,13 @@ namespace tfs
             end_key_buff = NULL;
           }
       }
-      /*
-      restful_client.logout();
-      */
-      clean_task_state_ = 0;
+
+      gettimeofday(&end, NULL);
+      int32_t timeuse = 1000000 * (end.tv_sec - start.tv_sec) + end.tv_usec - start.tv_usec;
+
+      TBSYS_LOG(INFO, "time: %d task use %d us, scan use %d us, unlink %d file", task_time, timeuse,
+          g_scan_time, expire_file_count);
+      atomic_dec(&running_task_count_);
       return ret;
     }//end for clean task
